[SUS-46]Refactor/social_login_addup_to_member
