+++ src/member/repository/MemberChangeRepository.ts
@@ -8,15 +8,6 @@ export default class MemberChangeRepository {
       id,
     ])
   }
-  /** 비밀번호 찾기 */
-  static async checkMemberPasswordFromDb(id: string, email: string) {
-    return (
-      await postgres.query(
-        'SELECT 1 FROM member WHERE id = $1 AND email = $2',
-        [id, email],
-      )
-    ).rowCount
-  }
 
   /** 비밀번호 변경
    * 비밀번호 값을 변경합니다. */
