+++ src/member/router/findRouter.ts
@@ -1,6 +1,7 @@
 import controller from '#controller'
 import express from 'express'
-import FindPasswordRequest from '../dao/frontend/request/FindPasswordRequest.js'
+import FindPasswordRequest from '../entity/dao/frontend/request/FindPasswordRequest.js'
+import FindService from '../service/FindService.js'
 export const findRouter = express.Router()
 
 /** 비밀번호 찾기 API */
@@ -12,7 +13,7 @@ findRouter.post(
     FindPasswordRequest,
     null,
   )(async (req, res): Promise<any> => {
-    await MemberService.findPassword(req.body)
+    await FindService.findPassword(req.body)
     return res.sendStatus(200)
   }),
 )
