+++ src/member/service/ChangeService.ts
@@ -2,9 +2,9 @@ import ErrorRegistry from '#error/ErrorRegistry'
 import EmailSender from '#util/email/mailSender/index'
 import Token from '#util/token/index'
 import { Request } from 'express'
-import NicknameChangeRequest from '../dao/frontend/request/NicknameChangeRequest.js'
-import PasswordChangeRequest from '../dao/frontend/request/PasswordChangeRequest.js'
-import SendFindPasswordEmailRequest from '../dao/frontend/request/SendFindPasswordEmailRequest.js'
+import NicknameChangeRequest from '../entity/dao/frontend/request/NicknameChangeRequest.js'
+import PasswordChangeRequest from '../entity/dao/frontend/request/PasswordChangeRequest.js'
+import SendFindPasswordEmailRequest from '../entity/dao/frontend/request/SendFindPasswordEmailRequest.js'
 import MemberChangeRepository from '../repository/MemberChangeRepository.js'
 import RedisEmailChangeRepository from '../repository/RedisEmailChangeRepository.js'
 export default class ChangeService {
